Move theme object into own file inside Styles folder
